feat: create PDFGenerator.swift and move pdf generator functions into the file
