+++ InvoiceManager/.swiftlint.yml
@@ -3,6 +3,7 @@ disabled_rules:
     - multiple_closures_with_trailing_closure
     - identifier_name
     - force_cast
+    - trailing_whitespace
 
 excluded:
     - Pods
